Inline script_pubkey ScriptBuf extension trait
The separation of extension traits into the `script_pubkey` module
does not assist clarity and makes it less ergonomic for users having
to import traits from different locations. We can put these methods on
the `script::ScriptBufExt` and fix both issues.

Once this is done we can remove the `script_pubkey` module entirely.
