+++ bitcoin/examples/ecdsa-psbt.rs
@@ -31,7 +31,6 @@
 use std::collections::BTreeMap;
 use std::fmt;
 
-use bitcoin::address::script_pubkey::ScriptBufExt as _;
 use bitcoin::bip32::{ChildNumber, DerivationPath, Fingerprint, IntoDerivationPath, Xpriv, Xpub};
 use bitcoin::consensus::encode;
 use bitcoin::consensus_validation::TransactionExt as _;
