+++ bitcoin/examples/taproot-psbt.rs
@@ -77,7 +77,6 @@ const UTXO_3: P2trUtxo = P2trUtxo {
 
 use std::collections::BTreeMap;
 
-use bitcoin::address::script_pubkey::ScriptBufExt as _;
 use bitcoin::bip32::{ChildNumber, DerivationPath, Fingerprint, Xpriv, Xpub};
 use bitcoin::consensus::encode;
 use bitcoin::consensus_validation::TransactionExt as _;
