+++ bitcoin/src/blockdata/script/owned.rs
@@ -5,12 +5,20 @@ use core::ops::Deref;
 
 use hex::FromHex as _;
 use internals::ToU64 as _;
+use secp256k1::{Secp256k1, Verification};
 
 use super::{opcode_to_verify, Builder, Instruction, PushBytes, ScriptExtPriv as _};
 use crate::consensus;
+use crate::key::{
+    PubkeyHash, PublicKey, TapTweak, TweakedPublicKey, UntweakedPublicKey, WPubkeyHash,
+};
 use crate::opcodes::all::*;
 use crate::opcodes::{self, Opcode};
 use crate::prelude::Vec;
+use crate::script::witness_program::{WitnessProgram, P2A_PROGRAM};
+use crate::script::witness_version::WitnessVersion;
+use crate::script::{self, ScriptHash, WScriptHash};
+use crate::taproot::TapNodeHash;
 
 #[rustfmt::skip]            // Keep public re-exports separate.
 #[doc(inline)]
@@ -25,6 +33,75 @@ crate::internal_macros::define_extension_trait! {
         /// Generates OP_RETURN-type of scriptPubkey for the given data.
         fn new_op_return<T: AsRef<PushBytes>>(data: T) -> Self {
             Builder::new().push_opcode(OP_RETURN).push_slice(data).into_script()
+        }
+
+        /// Generates P2PK-type of scriptPubkey.
+        fn new_p2pk(pubkey: PublicKey) -> Self {
+            Builder::new().push_key(pubkey).push_opcode(OP_CHECKSIG).into_script()
+        }
+
+        /// Generates P2PKH-type of scriptPubkey.
+        fn new_p2pkh(pubkey_hash: PubkeyHash) -> Self {
+            Builder::new()
+                .push_opcode(OP_DUP)
+                .push_opcode(OP_HASH160)
+                .push_slice(pubkey_hash)
+                .push_opcode(OP_EQUALVERIFY)
+                .push_opcode(OP_CHECKSIG)
+                .into_script()
+        }
+
+        /// Generates P2SH-type of scriptPubkey with a given hash of the redeem script.
+        fn new_p2sh(script_hash: ScriptHash) -> Self {
+            Builder::new()
+                .push_opcode(OP_HASH160)
+                .push_slice(script_hash)
+                .push_opcode(OP_EQUAL)
+                .into_script()
+        }
+
+        /// Generates P2WPKH-type of scriptPubkey.
+        fn new_p2wpkh(pubkey_hash: WPubkeyHash) -> Self {
+            // pubkey hash is 20 bytes long, so it's safe to use `new_witness_program_unchecked` (Segwitv0)
+            script::new_witness_program_unchecked(WitnessVersion::V0, pubkey_hash)
+        }
+
+        /// Generates P2WSH-type of scriptPubkey with a given hash of the redeem script.
+        fn new_p2wsh(script_hash: WScriptHash) -> Self {
+            // script hash is 32 bytes long, so it's safe to use `new_witness_program_unchecked` (Segwitv0)
+            script::new_witness_program_unchecked(WitnessVersion::V0, script_hash)
+        }
+
+        /// Generates P2TR for script spending path using an internal public key and some optional
+        /// script tree Merkle root.
+        fn new_p2tr<C: Verification, K: Into<UntweakedPublicKey>>(
+            secp: &Secp256k1<C>,
+            internal_key: K,
+            merkle_root: Option<TapNodeHash>,
+        ) -> Self {
+            let internal_key = internal_key.into();
+            let (output_key, _) = internal_key.tap_tweak(secp, merkle_root);
+            // output key is 32 bytes long, so it's safe to use `new_witness_program_unchecked` (Segwitv1)
+            script::new_witness_program_unchecked(WitnessVersion::V1, output_key.serialize())
+        }
+
+        /// Generates P2TR for key spending path for a known [`TweakedPublicKey`].
+        fn new_p2tr_tweaked(output_key: TweakedPublicKey) -> Self {
+            // output key is 32 bytes long, so it's safe to use `new_witness_program_unchecked` (Segwitv1)
+            script::new_witness_program_unchecked(WitnessVersion::V1, output_key.serialize())
+        }
+
+        /// Generates pay to anchor output.
+        fn new_p2a() -> Self {
+            script::new_witness_program_unchecked(WitnessVersion::V1, P2A_PROGRAM)
+        }
+
+        /// Generates P2WSH-type of scriptPubkey with a given [`WitnessProgram`].
+        fn new_witness_program(witness_program: &WitnessProgram) -> Self {
+            Builder::new()
+                .push_opcode(witness_program.version().into())
+                .push_slice(witness_program.program())
+                .into_script()
         }
 
         /// Constructs a new [`ScriptBuf`] from a hex string.
