+++ bitcoin/src/blockdata/script/tests.rs
@@ -3,9 +3,10 @@
 use hex_lit::hex;
 
 use super::*;
-use crate::address::script_pubkey::ScriptBufExt as _;
 use crate::consensus::encode::{deserialize, serialize};
 use crate::crypto::key::{PublicKey, XOnlyPublicKey};
+use crate::script::witness_program::WitnessProgram;
+use crate::script::witness_version::WitnessVersion;
 use crate::{opcodes, Amount, FeeRate};
 
 #[test]
@@ -982,4 +983,26 @@ fn instruction_script_num_parse() {
 fn script_push_int_overflow() {
     // Only errors if `data == i32::MIN` (CScriptNum cannot have value -2^31).
     assert_eq!(Builder::new().push_int(i32::MIN), Err(Error::NumericOverflow));
+}
+
+#[test]
+fn shortest_witness_program() {
+    let bytes = [0x00; 2]; // Arbitrary bytes, witprog must be between 2 and 40.
+    let version = WitnessVersion::V15; // Arbitrary version number, intentionally not 0 or 1.
+
+    let p = WitnessProgram::new(version, &bytes).expect("failed to create witness program");
+    let script = ScriptBuf::new_witness_program(&p);
+
+    assert_eq!(script.witness_version(), Some(version));
+}
+
+#[test]
+fn longest_witness_program() {
+    let bytes = [0x00; 40]; // Arbitrary bytes, witprog must be between 2 and 40.
+    let version = WitnessVersion::V16; // Arbitrary version number, intentionally not 0 or 1.
+
+    let p = WitnessProgram::new(version, &bytes).expect("failed to create witness program");
+    let script = ScriptBuf::new_witness_program(&p);
+
+    assert_eq!(script.witness_version(), Some(version));
 }
